完成group home 页面
